Automatically backed up by Learn

+++ app/controllers/tweets_controller.rb
@@ -1,43 +1,71 @@
 class TweetsController < ApplicationController
   get '/tweets' do
     if logged_in?
-      @user = User.find(session[:user_id])
-      erb :'tweets/index'
+      @tweets = Tweet.all
+      @user = current_user
+
+      erb :'tweets/tweets'
     else
       redirect '/login'
     end
   end
 
   get '/tweets/new' do
-
-    erb :'tweets/new'
+    if logged_in?
+      erb :'tweets/new'
+    else
+      redirect '/login'
+    end
   end
 
   get '/tweets/:id' do
-    @tweet = Tweet.find(params[:id])
+    if logged_in?
+      @tweet = Tweet.find(params[:id])
 
-    erb :'tweets/show'
+      erb :'tweets/show_tweet'
+    else
+      redirect '/login'
+    end
   end
 
   get '/tweets/:id/edit' do
-    @tweet = Tweet.find(params[:id])
+    if logged_in?
+      @tweet = Tweet.find(params[:id])
 
-    erb :'tweets/edit'
+      erb :'tweets/edit_tweet'
+    else
+      redirect '/login'
+    end
   end
 
   post '/tweets' do
-    erb :'tweets/index'
+    if params[:content] != ""
+      tweet = Tweet.create(:content => params[:content])
+      current_user.tweets << tweet
+      @user = current_user
+      @tweets = Tweet.all
+      erb :'tweets/tweets'
+    else
+      redirect 'tweets/new'
+    end
   end
 
   post '/tweets/:id/delete' do
     tweet = Tweet.find(params[:id])
-    tweet.destroy
+    if current_user.tweets.include?(tweet)
+      tweet.destroy
+    end
+    redirect '/tweets'
   end
 
-  patch '/tweets' do
-    tweet = Tweet.find(params[:id])
-    tweet.content = params[:tweet][:content]
-
-    redirect "/tweets/#{tweet.id}"
+  patch '/tweets/:id' do
+    if params[:content] != ""
+      tweet = Tweet.find(params[:id])
+      tweet.content = params[:content]
+      tweet.save
+      redirect "/tweets/#{tweet.id}"
+    else
+      redirect "/tweets/#{params[:id]}/edit"
+    end
   end
 end
